feat(data utils): вынести функции обработки в data_utils.py
вынести функции обработки в data_utils.py
