+++ src/data/convert_labels_to_CVAT_json.py
@@ -1,16 +1,10 @@
 import json
 
-
-def read_json(filename: str):
-    with open(filename) as f_in:
-        dct_data = json.load(f_in)
-
-    return dct_data
-
+from data_utils import read_json
 
 if __name__ == "__main__":
-    input_rtsd_json_labels_path = "data/RTSD/label_map.json"
-    output_cvat_json_path = "data/RSTD_cvat_val/label.json"
+    input_rtsd_json_labels_path = "data/interim/RTSD/label_map.json"
+    output_cvat_json_path = "data/interim/RTSD_cvat_train/label.json"
 
     label_to_num_mapping = read_json(input_rtsd_json_labels_path)
